Make PostOffice more idiomatic to use as a library
